Fix tiny scrollbar not being draggable.

+++ src/main/java/net/blay09/mods/cookingbook/client/GuiRecipeBook.java
@@ -167,7 +167,7 @@ public class GuiRecipeBook extends GuiContainer {
 		drawTexturedModalRect(x, y, 0, 0, xSize, ySize);
 
 		if (mouseClickY != -1) {
-			float pixelsPerFilter = (SCROLLBAR_HEIGHT - scrollBarScaledHeight) / Math.max(1, (int) Math.ceil(container.getAvailableRecipeCount() / 3f) - VISIBLE_ROWS);
+			float pixelsPerFilter = (SCROLLBAR_HEIGHT - scrollBarScaledHeight) / (float) Math.max(1, (int) Math.ceil(container.getAvailableRecipeCount() / 3f) - VISIBLE_ROWS);
 			if (pixelsPerFilter != 0) {
 				int numberOfFiltersMoved = (int) ((mouseY - mouseClickY) / pixelsPerFilter);
 				if (numberOfFiltersMoved != lastNumberOfMoves) {
